init function rename
